requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.4
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
drf-nested-routers==0.92.5
gunicorn==20.0.4
importlib-metadata==3.3.0
Markdown==3.3.3
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
